BUGFIX of the ':01' BUG? (Needs testing)
The bug was in the method to get the last errors.
The method was making a select in the DB for the MAX(exec_time) and that
causes a problem, because it will only select the ':00' or ':01', never
both together.

Changed the method of which gets the last exec_time to returns the time
without the seconds and the method that select all last errors to use a
'LIKE' instead of '='